requirements.txt
environs==9.5.0
requests==2.27.1
python-telegram-bot==13.11
flake8==5.0.4